protosvit/pyproject.toml
[project]
name = "protos-vit"
version = "0.1.0"
description = "protos-vit implementation"
dependencies = [
    "pytorch-lightning>=2.0.9.post0",
    "torchmetrics>=1.2.0",
    "numpy>=1.25.2",
    "pandas>=2.1.0",
    "hydra-core>=1.3.2",
    "pyrootutils>=1.0.4",
    "setuptools>=68.2.2",
    "rich>=13.6.0",
    "timm>=0.9.7",
    "tensorboard>=2.14.1",
    "ipykernel>=6.25.2",
    "matplotlib>=3.8.0",
    "transformers>=4.34.0",
    "opencv-python>=4.8.1.78",
    "wandb>=0.15.12",
    "scikit-learn>=1.3.1",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "ipywidgets>=8.1.1",
    "pydicom>=2.4.3",
    "einops>=0.7.0",
    "plotly>=5.18.0",
    "pre-commit>=3.6.0",
    "open-clip-torch>=2.24.0",
    "gdown>=5.2.0",
]
requires-python = ">=3.10"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
dev = ["seaborn>=0.13.1"]

[tool.pdm.scripts]
train_classification = "src/main_train.py experiment=classification_prototype"
launch_tensorboard = { call = "src.shared_utils.tensorboard_launcher:main" }


[tool.ruff]
target-version = "py312"
extend-include = ["*.ipynb"]
select = ["ALL"]
ignore = [
    "FA",     # ignore import future declaration
    "F401",   # unused-import
    "ANN201", # return types enforcing
    "T201",   # allow prints
    "D100",   # ignore missing docs
    "D101",   # ^
    "D102",   # ^
    "D103",   # ^
    "D104",   # ^
    "D105",   # ^
    "D106",   # ^
    "D107",   # ^
    "D200",   # ignore docs rules
    "D205",   # ^
    "D212",   # ^
    "D400",   # ^
    "D401",   # ^
    "D415",   # ^
]
